requirement.txt
Aspose.Slides==22.7.0